Fixed Bunker research times and supply times
Supposedly, it's 8h for no upgrades and 3h 30min for 2 upgrades. With
all the info I got and some calculations, as it has to fill an entire
research bar to unlock an item, 100 points are needed with no upgrades
and 60 with 2. So 80 should be with only one of them, that makes it 8h
20min with no upgrades (confirmed), 7h 5min with one (calculated) and
5h 50min with two.

As for supplies I used the same method to calculate the times of the MC
business given the solid data I found to get 127.5min